+++ SConstruct
@@ -14,7 +14,7 @@ env = SConscript("godot-cpp/SConstruct")
 
 # tweak this if you want to use different folders, or more folders, to store your source code in.
 env.Append(CPPPATH=["extension/src/"])
-sources = Glob("extension/src/*.cpp")
+sources = Glob("extension/src/**/*.cpp")
 
 if env["platform"] == "macos":
     library = env.SharedLibrary(
