+++ extension/src/nodes/beehave_tree.cpp
@@ -1,12 +1,11 @@
-#include "tree.h"
-
+#include "beehave_tree.h"
 #include <godot_cpp/core/class_db.hpp>
 
 using namespace godot;
 
 BeehaveTree::BeehaveTree()
 {
-    count = 0;
+
 }
 
 BeehaveTree::~BeehaveTree()
@@ -14,7 +13,11 @@ BeehaveTree::~BeehaveTree()
     
 }
 
+void BeehaveTree::test() {
+
+}
+
 void BeehaveTree::_bind_methods()
 {
-
+    ClassDB::bind_method(D_METHOD("test"), &BeehaveTree::test);
 }
