+++ extension/src/nodes/beehave_tree_node.cpp
@@ -1,0 +1,23 @@
+#include "beehave_tree_node.h"
+#include <godot_cpp/core/class_db.hpp>
+
+using namespace godot;
+
+BeehaveTreeNode::BeehaveTreeNode()
+{
+
+}
+
+BeehaveTreeNode::~BeehaveTreeNode()
+{
+    
+}
+
+void BeehaveTreeNode::test() {
+
+}
+
+void BeehaveTreeNode::_bind_methods()
+{
+    ClassDB::bind_method(D_METHOD("test"), &BeehaveTreeNode::test);
+}
