--- extension/src/nodes/blackboard.cpp
@@ -1,20 +1,0 @@
-#include "blackboard.h"
-
-#include <godot_cpp/core/class_db.hpp>
-
-using namespace godot;
-
-BeehaveBlackboard::BeehaveBlackboard()
-{
-    count = 0;
-}
-
-BeehaveBlackboard::~BeehaveBlackboard()
-{
-    
-}
-
-void BeehaveBlackboard::_bind_methods()
-{
-
-}
