--- extension/src/nodes/node.cpp
@@ -1,20 +1,0 @@
-#include "node.h"
-
-#include <godot_cpp/core/class_db.hpp>
-
-using namespace godot;
-
-BeehaveNode::BeehaveNode()
-{
-
-}
-
-BeehaveNode::~BeehaveNode()
-{
-    
-}
-
-void BeehaveNode::_bind_methods()
-{
-
-}
